Fixed IDLE in library
The couples were selected from choosen to be killed phenotypes.
Problem was because 'diff' was invoked on collections containing different types ActorRef and Evaluated

I made random killer optional to be sure it not makes any error

Second error was with GodFathers phenotypesToEvaluate. It was not initialized correctly so there was wrong value. Now it's only counter